better click regions
The main improvement of this patch over the previous version 20210405 is that
the click region now ends on a matching signal raw byte.

The matching byte is optional, and without it dwm will behave as before.

To take advantage of this feature, scripts need to be modified to print the raw
byte at the end as well.

In addition, this patch cleanly applies onto master branch.